move individual tag test to separate files
